+++ routes/home.py
@@ -1,12 +1,15 @@
 # -*- coding: utf-8 -*-
 from flask import request, Blueprint
 
+from service.auth import auth_user
 from service.bangumi import bangumi_service
 from service.watch import watch_service
 from service.announce import announce_service
 from flask_login import login_required, current_user
 from domain.Favorites import Favorites
+import json
 
+from utils.exceptions import ClientError
 
 home_api = Blueprint('home', __name__)
 
@@ -21,11 +24,12 @@ def recent_update():
 @home_api.route('/on_air', methods=['GET'])
 @login_required
 def on_air_bangumi():
-    type = request.args.get('type', 2)
-    return bangumi_service.on_air_bangumi(current_user.id, type)
+    bangumi_type = request.args.get('type', 2)
+    return bangumi_service.on_air_bangumi(current_user.id, bangumi_type)
 
 
 @home_api.route('/my_bangumi', methods=['GET'])
+@auth_user(0)
 def my_bangumi():
     status = int(request.args.get('status', Favorites.WATCHING))
     if status == 0:
@@ -34,12 +38,14 @@ def my_bangumi():
 
 
 @home_api.route('/episode/<episode_id>', methods=['GET'])
+@auth_user(0)
 @login_required
 def episode_detail(episode_id):
     return bangumi_service.episode_detail(episode_id, current_user.id)
 
 
 @home_api.route('/bangumi', methods=['GET'])
+@auth_user(0)
 @login_required
 def list_bangumi():
     page = int(request.args.get('page', 1))
@@ -58,6 +64,7 @@ def list_bangumi():
 
 
 @home_api.route('/bangumi/<bangumi_id>', methods=['GET'])
+@auth_user(0)
 @login_required
 def bangumi_detail(bangumi_id):
     return bangumi_service.get_bangumi(bangumi_id, current_user.id)
@@ -67,3 +74,16 @@ def bangumi_detail(bangumi_id):
 @login_required
 def get_available_announce():
     return announce_service.get_available_announce()
+
+
+@home_api.route('/feedback', methods=['POST'])
+@auth_user(0)
+@login_required
+def feed_back():
+    data = json.loads(request.get_data(as_text=True))
+    episode_id = data.get('episode_id', None)
+    video_file_id = data.get('video_file_id', None)
+    message= data.get('message', None)
+    if not episode_id or not video_file_id:
+        raise ClientError(ClientError, ClientError.INVALID_REQUEST)
+    return bangumi_service.feed_back(episode_id, video_file_id, current_user, message)